+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   factorisation/renormalisation scales
 - `pineappl` has a new subcommand `pdf_uncertainty` to calculate PDF
   uncertainties
+- in `examples/capi-dy-aa` one can find an example how to use the C API
 
 ## [0.1.0] - 11/06/2020